Fix StandardHeader fields

+++ composeApp/src/desktopMain/kotlin/com/alekso/dltstudio/ui/logs/LazyScrollable.kt
@@ -54,7 +54,7 @@ fun LazyScrollable(
                                     }),
                                 i.toString(),
                                 simpleDateFormat.format(message.timeStampSec * 1000L + message.timeStampUs / 1000),
-                                if (message.standardHeader.timeStamp != null) "%.4f".format(message.standardHeader.timeStamp!! / 10000f) else "-",
+                                if (message.standardHeader.timeStamp != null) "%.4f".format(message.standardHeader.timeStamp!!.toLong() / 10000f) else "-",
                                 message.ecuId,
                                 "${message.standardHeader.ecuId}",
                                 "${message.standardHeader.sessionId}",
